requirements.txt
python-telegram-bot==21.4
ipython==8.27.0 # mainly to test coroutines in the shell
python_dotenv==1.0.1 # to ensure .env files are always loaded